Make mongo changes for multiple collection handling
