+++ config/config_test.go
@@ -30,8 +30,7 @@ func TestConfig(t *testing.T) {
 				So(cfg.DefaultLimit, ShouldEqual, 20)
 				So(cfg.DefaultMaxLimit, ShouldEqual, 100)
 				So(cfg.DefaultOffset, ShouldEqual, 0)
-				So(cfg.MongoConfig.BindAddr, ShouldEqual, "localhost:27017")
-				So(cfg.MongoConfig.Collection, ShouldEqual, "metadata")
+				So(cfg.MongoConfig.ClusterEndpoint, ShouldEqual, "localhost:27017")
 				So(cfg.MongoConfig.Database, ShouldEqual, "interactives")
 				So(cfg.MongoConfig.Username, ShouldEqual, "")
 				So(cfg.MongoConfig.Password, ShouldEqual, "")
