+++ features/getinteractive.feature
@@ -43,7 +43,7 @@ Feature: Interactives API (Get interactive)
                     },
                     "state": "ArchiveUploaded",
                     "last_updated":"2021-01-01T00:00:00Z",
-                    "url": "http://localhost:27300/interactives/slug-abcde123/embed",
+                    "url": "http://preview_url/interactives/slug-abcde123/embed",
                     "uri": "/interactives/slug-abcde123"
                 }
             """
