+++ features/listinteractivesfilter.feature
@@ -55,7 +55,7 @@ Feature: Interactives API (List interactives)
                                 "name": "rhyCq4GCknxx0nzeqx2LE077Ruo=/TestMe.zip"
                             },
                             "last_updated":"2021-01-01T00:00:01Z",
-                            "url": "http://localhost:27300/interactives/slug-abcde2/embed",
+                            "url": "http://preview_url/interactives/slug-abcde2/embed",
                             "uri": "/interactives/slug-abcde2"
                         }
                     ]
@@ -115,7 +115,7 @@ Feature: Interactives API (List interactives)
                                 "name": "rhyCq4GCknxx0nzeqx2LE077Ruo=/TestMe.zip"
                             },
                             "last_updated":"2021-01-01T00:00:01Z",
-                            "url": "http://localhost:27300/interactives/slug-resid2/embed",
+                            "url": "http://preview_url/interactives/slug-resid2/embed",
                             "uri": "/interactives/slug-resid2"
                         }
                     ]
@@ -178,7 +178,7 @@ Feature: Interactives API (List interactives)
                                 "name": "rhyCq4GCknxx0nzeqx2LE077Ruo=/TestMe.zip"
                             },
                             "last_updated":"2021-01-01T00:00:01Z",
-                            "url": "http://localhost:27300/interactives/slug-resid2/embed",
+                            "url": "http://preview_url/interactives/slug-resid2/embed",
                             "uri": "/interactives/slug-resid2"
                         }
                 ]
@@ -237,7 +237,7 @@ Feature: Interactives API (List interactives)
                                 "name": "kqA7qPo1GeOJeff69lByWLbPiZM=/docker-vernemq-master.zip"
                             },
                             "last_updated":"2021-01-01T00:00:00Z",
-                            "url": "http://localhost:27300/interactives/slug-resid1/embed",
+                            "url": "http://preview_url/interactives/slug-resid1/embed",
                             "uri": "/interactives/slug-resid1"
                         },
                         {
@@ -256,7 +256,7 @@ Feature: Interactives API (List interactives)
                                 "name": "rhyCq4GCknxx0nzeqx2LE077Ruo=/TestMe.zip"
                             },
                             "last_updated":"2021-01-01T00:00:01Z",
-                            "url": "http://localhost:27300/interactives/slug-resid2/embed",
+                            "url": "http://preview_url/interactives/slug-resid2/embed",
                             "uri": "/interactives/slug-resid2"
                         }
                     ]
